+++ sdk/release_notes.md
@@ -4,9 +4,15 @@
 - My change description (#PR/#issue)
 -->
 
-### Microsoft.Azure.Functions.Worker.Sdk 1.11.0 (meta package)
+### Microsoft.Azure.Functions.Worker.Sdk <version> (meta package)
+
 - Update extension build project to reference Microsoft.NET.Sdk.Functions 4.2.0
+
 ### Microsoft.Azure.Functions.Worker.Sdk.Analyzers <version> (delete if not updated)
+
 - <entry>
-### Microsoft.Azure.Functions.Worker.Sdk.Generators 1.1.0-preview3
-- <entry>
+
+### Microsoft.Azure.Functions.Worker.Sdk.Generators <version>
+
+- Add support for retry options (#1548)
+- Bug fix for when DefaultValue is not present on an IsBatched prop (#1602).